+++ backends/simplego/packgemm/gen_gemm16reg_impl_fallback.gen.go
@@ -225,7 +225,7 @@ func BaseGEMMSymmetric16RegistersGemmChunk_fallback_Float16(alpha hwy.Float16, b
 				if contractingPanelIdx > 0 {
 					effectiveBeta = 1
 				}
-				ApplyPackedOutput(packedOutput, output, alpha, hwy.Float32ToFloat16(effectiveBeta), params.RHSPanelCrossSize, lhsPanelRowIdx, rhsPanelColIdx, rhsCrossSize, lhsPanelHeight, rhsPanelWidth)
+				ApplyPackedOutput(packedOutput, output, alpha, hwy.Float32ToFloat16(float32(effectiveBeta)), params.RHSPanelCrossSize, lhsPanelRowIdx, rhsPanelColIdx, rhsCrossSize, lhsPanelHeight, rhsPanelWidth)
 			}
 		}
 	}
@@ -245,7 +245,7 @@ func BaseGEMMSymmetric16RegistersGemmChunk_fallback_BFloat16(alpha hwy.BFloat16,
 				if contractingPanelIdx > 0 {
 					effectiveBeta = 1
 				}
-				ApplyPackedOutput(packedOutput, output, alpha, hwy.Float32ToBFloat16(effectiveBeta), params.RHSPanelCrossSize, lhsPanelRowIdx, rhsPanelColIdx, rhsCrossSize, lhsPanelHeight, rhsPanelWidth)
+				ApplyPackedOutput(packedOutput, output, alpha, hwy.Float32ToBFloat16(float32(effectiveBeta)), params.RHSPanelCrossSize, lhsPanelRowIdx, rhsPanelColIdx, rhsCrossSize, lhsPanelHeight, rhsPanelWidth)
 			}
 		}
 	}
